Publish model in Fieldset.onRefreshRequirements callback
The verification-components consumer was propagating the onRefreshRequirements callback and relying on the their two way bound model to have been updated. But due to the sequence of events in Angular they were propagating the old model. By supplying the latest model as a parameter, we make it easier for consumers to send the correct data.