Broadcast an onModelChange with initial model value and validity on twFieldset load
Previously validity was communicated by the two-way bound isValid component prop. This was being misused by some consumers to attempt to set internal validity.  We now additionally broadcast validity in the onModelChange callback for a clearer API.

Previously it did not broadcast a change on load, meaning you had to rely on the two-way bound isValid property for initial validation state.  This also fixes a bug where we were not validating required fields.